backend/requirements.txt
asgiref==3.9.1
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djongo==1.3.7
gunicorn==23.0.0
packaging==25.0
PyJWT==2.10.1
pymongo==3.11.4
pytz==2025.2
sqlparse==0.5.3
whitenoise==6.11.0
